requirements/base.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
crispy-bootstrap3==2024.1 ; python_version >= "3.12" and python_version < "4.0"
crispy-bootstrap5==2025.4 ; python_version >= "3.12" and python_version < "4.0"
cryptography==44.0.2 ; python_version >= "3.12" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
django-configurations==2.5.1 ; python_version >= "3.12" and python_version < "4.0"
django-crispy-forms==2.3 ; python_version >= "3.12" and python_version < "4.0"
django-filter==25.1 ; python_version >= "3.12" and python_version < "4.0"
django-prometheus==2.3.1 ; python_version >= "3.12" and python_version < "4.0"
django==5.2 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework-simplejwt==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.16.0 ; python_version >= "3.12" and python_version < "4.0"
djoser==2.3.1 ; python_version >= "3.12" and python_version < "4.0"
drf-extensions==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
drf-redesign==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
drf-spectacular-sidecar==2025.4.1 ; python_version >= "3.12" and python_version < "4.0"
drf-spectacular==0.28.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.12" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.12" and python_version < "4.0"
loguru==0.7.3 ; python_version >= "3.12" and python_version < "4.0"
mysqlclient==2.2.7 ; python_version >= "3.12" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
prometheus-client==0.21.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pyjwt==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
pymysql==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
pytest-django==4.11.1 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.5 ; python_version >= "3.12" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
referencing==0.36.2 ; python_version >= "3.12" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rpds-py==0.24.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
social-auth-app-django==5.4.3 ; python_version >= "3.12" and python_version < "4.0"
social-auth-core==4.5.6 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.13.1 ; python_version >= "3.12" and python_version < "3.13"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
win32-setctime==1.2.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
